Arranged the server codes and routers.
