+++ controllers/fileLoaderController.js
@@ -1,4 +1,3 @@
-require("dotenv").config();
 const path = require("path");
 
 module.exports = {
@@ -6,4 +5,7 @@ module.exports = {
         res.sendFile(path.join(__dirname, '../loginAndForm/market.html'));
     },
 
+    loadMassageFile: (req, res) => {
+        res.sendFile(path.join(__dirname, '../loginAndForm/message.html'));
+    },
 }
