+++ controllers/subscriptionController.js
@@ -1,5 +1,5 @@
 const { Subscription } = require('../models/subscriptions')
 
-exports.subsController = {
+module.exports = {
 
 }
