+++ routers/planRouter.js
@@ -1,7 +1,9 @@
 const { Router } = require('express');
-const {planController} = require('../controllers/planControllerController');
+const planController = require('../controllers/planController');
+const planRouter = Router();
 
-const planRouter = new Router();
+planRouter.get('/', planController.getAllPlans)
+planRouter.post('/', planController.purchasePlan)
 
 
-module.exports = {planRouter};
+module.exports = planRouter;
